Refactor Tools to write to files instead of the database
Tools & Reports were writing to the database which caused performance issues.
Instead of writing to the database and storing large content there, we now
write files into the "protected" file storage and serve it currently via
the application itself. In the future we can use Nginx to handle the file
directly, however this was not done at this time.