fix: Number creation could raise an error
If a number was previously created and then shadow deleted
(active=False), using the API endpoint for POST would raise an
exception. It is recommended to clarify the operations such that a
DELETE operation on the Number table actually deletes. This will make
the API consistent. Using a PUT to perform a create is not industry
practice or clear.